… pytensor/tensor/rewriting/blas.py
… pytensor/tensor/rewriting/blockwise.py
… pytensor/tensor/rewriting/linalg.py
… pytensor/tensor/rewriting/subtensor.py
… pytensor/tensor/shape.py
… pytensor/tensor/subtensor.py
… tests/link/jax/test_nlinalg.py
… tests/link/numba/test_basic.py
… tests/tensor/rewriting/test_basic.py
… tests/tensor/rewriting/test_blas.py
… tests/tensor/rewriting/test_blockwise.py
… tests/tensor/rewriting/test_subtensor.py
… tests/tensor/test_basic.py
… tests/tensor/test_blockwise.py
… tests/tensor/test_subtensor.py