This PR adds a bunch of blockwise improvements, using `test_batched_mvnormal_logp_and_dlogp` as a benchmark.

## TODO:
- [x] Direct test for Blockwise Alloc rewrite
- [x] Direct test for Blockwise AdvancedIncSubtensor rewrite

```
BEFORE:
------------------------------------------------------------------------------------------------------------------------- benchmark: 9 tests -------------------------------------------------------------------------------------------------------------------------
Name (time in us)                                                             Min                     Max                    Mean                 StdDev                  Median                     IQR            Outliers         OPS            Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_batched_mvnormal_logp_and_dlogp[cov:()-mu:()]                       241.4050 (1.0)        1,810.0030 (1.0)          267.8159 (1.0)         112.2918 (1.0)          248.6980 (1.0)            6.1395 (1.0)        43;202  3,733.9078 (1.0)        1725           1
test_batched_mvnormal_logp_and_dlogp[cov:()-mu:(1000,)]                7,727.6810 (32.01)     32,625.9060 (18.03)      9,140.6252 (34.13)     3,506.5375 (31.23)      8,056.3195 (32.39)        532.3450 (86.71)        8;14    109.4017 (0.03)        106           1
test_batched_mvnormal_logp_and_dlogp[cov:()-mu:(4, 1000)]             36,519.7910 (151.28)    43,507.6490 (24.04)     37,502.4438 (140.03)    1,357.3882 (12.09)     37,078.4155 (149.09)       775.1030 (126.25)        1;1     26.6649 (0.01)         26           1
test_batched_mvnormal_logp_and_dlogp[cov:(1000,)-mu:()]              204,345.6680 (846.48)   213,811.9620 (118.13)   209,570.1286 (782.52)    3,860.6208 (34.38)    211,285.3960 (849.57)     5,903.3477 (961.54)        2;0      4.7717 (0.00)          5           1
test_batched_mvnormal_logp_and_dlogp[cov:(1000,)-mu:(1000,)]         206,162.0440 (854.01)   246,957.8760 (136.44)   220,043.3346 (821.62)   15,613.3438 (139.04)   215,733.2810 (867.45)    11,453.7725 (>1000.0)       1;1      4.5446 (0.00)          5           1
test_batched_mvnormal_logp_and_dlogp[cov:(1000,)-mu:(4, 1000)]       344,049.8250 (>1000.0)  401,401.8120 (221.77)   371,411.4898 (>1000.0)  23,642.7591 (210.55)   364,708.9210 (>1000.0)   38,862.3615 (>1000.0)       2;0      2.6924 (0.00)          5           1
test_batched_mvnormal_logp_and_dlogp[cov:(4, 1000)-mu:()]            791,623.1920 (>1000.0)  937,836.2080 (518.14)   842,729.1312 (>1000.0)  68,029.4537 (605.83)   798,030.7500 (>1000.0)  110,470.4235 (>1000.0)       1;0      1.1866 (0.00)          5           1
test_batched_mvnormal_logp_and_dlogp[cov:(4, 1000)-mu:(4, 1000)]     806,370.6470 (>1000.0)  988,275.6840 (546.01)   896,315.0546 (>1000.0)  79,760.8230 (710.30)   928,718.6800 (>1000.0)  135,838.3053 (>1000.0)       2;0      1.1157 (0.00)          5           1
test_batched_mvnormal_logp_and_dlogp[cov:(4, 1000)-mu:(1000,)]       820,059.1180 (>1000.0)  877,132.3050 (484.60)   848,327.1440 (>1000.0)  20,863.3722 (185.80)   844,725.1540 (>1000.0)   24,586.2695 (>1000.0)       2;0      1.1788 (0.00)          5           1
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

AFTER:
------------------------------------------------------------------------------------------------------------------------- benchmark: 9 tests ------------------------------------------------------------------------------------------------------------------------
Name (time in us)                                                             Min                     Max                    Mean                 StdDev                  Median                    IQR            Outliers         OPS            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_batched_mvnormal_logp_and_dlogp[cov:()-mu:()]                       223.7400 (1.0)        1,970.9100 (1.0)          239.0524 (1.0)          76.9007 (1.0)          229.3510 (1.0)           3.1460 (1.0)        35;196  4,183.1832 (1.0)        1854           1
test_batched_mvnormal_logp_and_dlogp[cov:()-mu:(1000,)]                  807.4360 (3.61)      19,562.5290 (9.93)         991.2214 (4.15)      1,043.0862 (13.56)        825.7610 (3.60)         22.8352 (7.26)       22;122  1,008.8563 (0.24)        779           1
test_batched_mvnormal_logp_and_dlogp[cov:()-mu:(4, 1000)]              2,792.7740 (12.48)     11,729.2920 (5.95)       3,331.1094 (13.93)       837.6202 (10.89)      3,153.6910 (13.75)       649.0978 (206.32)       14;8    300.2003 (0.07)        225           1
test_batched_mvnormal_logp_and_dlogp[cov:(1000,)-mu:(1000,)]         128,476.3110 (574.22)   131,474.4790 (66.71)    130,098.2186 (544.22)    1,087.8881 (14.15)    129,534.9570 (564.79)    1,687.7140 (536.46)        3;0      7.6865 (0.00)          9           1
test_batched_mvnormal_logp_and_dlogp[cov:(1000,)-mu:()]              128,885.2480 (576.05)   134,904.6290 (68.45)    131,217.4634 (548.91)    2,093.2579 (27.22)    130,645.1510 (569.63)    3,462.7852 (>1000.0)       3;0      7.6209 (0.00)          9           1
test_batched_mvnormal_logp_and_dlogp[cov:(1000,)-mu:(4, 1000)]       253,046.7460 (>1000.0)  267,180.3140 (135.56)   261,121.1530 (>1000.0)   6,518.4257 (84.76)    264,305.8970 (>1000.0)  11,556.8890 (>1000.0)       1;0      3.8296 (0.00)          5           1
test_batched_mvnormal_logp_and_dlogp[cov:(4, 1000)-mu:(1000,)]       478,298.3400 (>1000.0)  539,055.1810 (273.51)   507,703.0552 (>1000.0)  25,177.5062 (327.40)   503,584.5180 (>1000.0)  42,617.3770 (>1000.0)       2;0      1.9697 (0.00)          5           1
test_batched_mvnormal_logp_and_dlogp[cov:(4, 1000)-mu:()]            478,898.5630 (>1000.0)  488,528.9760 (247.87)   482,516.0360 (>1000.0)   3,638.3298 (47.31)    481,567.2750 (>1000.0)   3,815.3637 (>1000.0)       1;0      2.0725 (0.00)          5           1
test_batched_mvnormal_logp_and_dlogp[cov:(4, 1000)-mu:(4, 1000)]     479,454.1990 (>1000.0)  576,023.7790 (292.26)   513,581.6604 (>1000.0)  36,769.4058 (478.14)   501,011.6260 (>1000.0)  32,514.4020 (>1000.0)       1;0      1.9471 (0.00)          5           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

It also closes https://github.com/pymc-devs/pymc/issues/7042
In my machine VI sampling goes down from 5m35s to 53s