@ricardoV94 2023-12-09
@jessegrabowski the biggest speedups come from:

1. Avoiding Blockwise all together which doesn't have C implementation (nor numba, although that was not the driving reason) by implementing eager vectorize or, when not possible, specialization rewrites into other equivalent graphs like matmul to batched_dot.

2. Avoiding repeated computations on the Blockwise by pushing Alloc inputs to the outputs of Blockwise (think about doing `foo(alloc(x, 1000))` vs `alloc(foo(x), 100)` with the extra benefit that sometimes these allocs are not needed at all. This happens when there would be a subtensor operation later on, or they would be implicitly broadcasted by another input to an Elemwise or Blockwise that actually is "full rank".